Update Magic Wormhole to 0.12.0
Summary: Updated Magic Wormhole to 0.12.0. Which fixes the "requires ip-address package" error with 0.10.5 in the Solus repository. For the life of me I cannot find a changelog for this version. This update requires D8771 as a dependency.

Test Plan: Tested by sending and recieving a file between two computers. Everything functioned as expected.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D8772